pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matrix-maintainer"
version = "0.1.0"
description = "Daily health and standards maintainer for Agent-Matrix repositories."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "Apache-2.0"}
authors = [{name = "Agent-Matrix"}]
dependencies = [
  "PyGithub>=2.3.0",
  "GitPython>=3.1.43",
  "Jinja2>=3.1.4",
  "PyYAML>=6.0.2",
  "httpx>=0.27.0",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "tenacity>=8.5.0",
  "packaging>=24.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
]

[project.scripts]
matrix-maintainer = "matrix_maintainer.cli:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["matrix_maintainer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.uv]
dev-dependencies = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
]
